refactor: update KafkaUtils import paths and configuration for producer
