+++ pom.xml
@@ -37,7 +37,7 @@
     <openapi-tools.jackson-databind-nullable.version>0.2.10</openapi-tools.jackson-databind-nullable.version>
     <cql2pgjson.version>36.0.0</cql2pgjson.version>
     <folio-spring-cql.version>10.0.0</folio-spring-cql.version>
-    <applications-poc-tools.version>4.0.0-SNAPSHOT</applications-poc-tools.version>
+    <applications-poc-tools.version>4.1.0-SNAPSHOT</applications-poc-tools.version>
     <folio-java-checkstyle.version>1.2.0</folio-java-checkstyle.version>
     <semver4j.version>5.8.0</semver4j.version>
     <am.yaml-file>${project.basedir}/src/main/resources/swagger.api/am.yaml</am.yaml-file>
@@ -196,7 +196,7 @@
 
     <dependency>
       <groupId>org.folio</groupId>
-      <artifactId>folio-integration-kafka</artifactId>
+      <artifactId>folio-kafka-producer</artifactId>
       <version>${applications-poc-tools.version}</version>
     </dependency>
